babel: split stylesheet into smaller chunks
